requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
coloredlogs==15.0.1
humanfriendly==10.0
inflect==7.5.0
pyreadline3==3.5.4
PyYAML==6.0.3
requests==2.34.1
ruamel.yaml.clib==0.2.15
ruamel.yaml==0.19.1
urllib3==2.7.0
yamale==6.1.0
APScheduler~=3.10.4
Flask~=2.3.3
jsonschema~=4.17.3
Flask-APScheduler==1.13.1
dataclasses_json==0.6.7